New implementation of adpcm decoder for amv variant which is (hopefully) correct, as far as we can deduce from data in binary format itself.
Because of that, it implements starving decoder which will stop decoding in
middle of byte, but correct according to number of samples in frame.

AMV specific modification inside CODEC_ID_ADPCM_IMA_SMJPEG will have to be
removed if this encoder gets prefered over older one to ease merge into
ffmpeg.